Refactor accounts and users. Add notification proto definition.
