+++ api/api.proto
@@ -166,6 +166,11 @@ service Nakama {
     option (google.api.http).delete = "/v2/friend";
   }
 
+  // Delete one or more users by ID or username.
+  rpc DeleteNotifications (DeleteNotificationsRequest) returns (google.protobuf.Empty) {
+    option (google.api.http).delete = "/v2/notification";
+  }
+
   // Fetch the current user's account.
   rpc GetAccount (google.protobuf.Empty) returns (Account) {
     option (google.api.http).get = "/v2/account";
@@ -248,6 +253,11 @@ service Nakama {
   // List all friends for the current user.
   rpc ListFriends (google.protobuf.Empty) returns (Friends) {
     option (google.api.http).get = "/v2/friend";
+  }
+
+  // Fetch list of notifications.
+  rpc ListNotifications (ListNotificationsRequest) returns (NotificationList) {
+    option (google.api.http).get = "/v2/notification";
   }
 
   // Execute a Lua function on the server.
@@ -535,6 +545,12 @@ message DeleteFriendsRequest {
   repeated string usernames = 2;
 }
 
+// Delete one or more notifications for the current user.
+message DeleteNotificationsRequest {
+  // The id of notifications.
+  repeated string ids = 1;
+}
+
 // A friend of a user.
 message Friend {
   // The friendship status.
@@ -555,6 +571,12 @@ message Friend {
   User user = 1;
   // The friend status.
   int32 state = 2; // one of "Friend.State".
+}
+
+// A collection of zero or more friends of the user.
+message Friends {
+  // The Friend objects.
+  repeated Friend friends = 1;
 }
 
 // Fetch a batch of zero or more users from the server.
@@ -617,10 +639,38 @@ message LinkFacebookRequest {
   google.protobuf.BoolValue import = 4;
 }
 
-// A collection of zero or more friends of the user.
-message Friends {
-  // The Friend objects.
-  repeated Friend friends = 1;
+// Fetch a list of unexpired notifications.
+message ListNotificationsRequest {
+  // Max number of notifications to list. Between 1 and 100.
+  int64 limit = 1;
+  // Use this cursor to paginate notifications. Cache this to catch up to new notifications. The value of this comes from NotificationList.cachable_cursor.
+  string cachable_cursor = 2;
+}
+
+// A notification in the server.
+message Notification {
+  // ID of the Notification.
+  string id = 1;
+  // Subject of the notification.
+  string subject = 2;
+  // Content of the notification in JSON.
+  string content = 3;
+  // Category code for this notification.
+  int64 code = 4;
+  // ID of the sender, if a user. Otherwise 'null'.
+  string sender_id = 5;
+  // The UNIX time when the notification was created.
+  google.protobuf.Timestamp create_time = 6;
+  // True if this notification was persisted to the database.
+  bool persistent = 7;
+}
+
+// A collection of zero or more notifications.
+message NotificationList {
+  // Collection of notifications.
+  repeated Notification notifications = 1;
+  // Use this cursor to paginate notifications. Cache this to catch up to new notifications.
+  string cachable_cursor = 2;
 }
 
 // Execute an Lua function on the server.
